perf: use int8 dtype for OneHotEncoder in both code paths
Switch from float64 (8 bytes/nnz) to int8 (1 byte/nnz) for the OHE
sparse matrix. Uses int8 instead of bool to preserve numeric sparse
matrix semantics (sum, matmul work correctly without dtype surprises).

Applied to both the default OneHotEncoder path and CappedOneHotEncoder.

+++ sliceline/slicefinder.py
@@ -789,7 +789,9 @@ class Slicefinder(BaseEstimator, TransformerMixin):
                 max_cardinality=self.max_cardinality
             )
         else:
-            self._one_hot_encoder = OneHotEncoder(handle_unknown="ignore")
+            self._one_hot_encoder = OneHotEncoder(
+                handle_unknown="ignore", dtype=np.int8
+            )
         x_encoded = self._one_hot_encoder.fit_transform(input_x)
         feature_domains: NDArray = np.array(
             [len(sub_array) for sub_array in self._one_hot_encoder.categories_]
